Collected all Settlements.
Conversion of countries to mollweide to facilitate a poly2grid() in the grids. Broken up the countries definition with intermediate storage to fss

Creation of landcells domain from grids (all grids not null and within country borders). Broken up with intermediate storage to fss

Extended continental network generation. The prior scripts were insufficient, as (eg in Oceania) continents can consist of many disjointed networks that need to be maintained. Solution:
- Split countries poly into separate polygons (through split), identifies 'islands' big and small. This is the seldomain.
- Check the largest networks that have at least one node within the seldomain
- Select all the largest networks.

-> the current choice is to select the largest network with at least one node within the border. Even if there may locally be networks that have more nodes within borders. The preference is to retain a limited number of big networks, that ideally have cross-border connectivity.